// backend/src/common/enums/error-code.enum.ts
export enum ErrorCode {
  ORDER_NOT_FOUND = 'ORDER_NOT_FOUND',
  UNAUTHORIZED_ORDER = 'UNAUTHORIZED_ORDER',
  SUBSCRIPTION_PAUSED = 'SUBSCRIPTION_PAUSED',
  EDIT_CUTOFF_PASSED = 'EDIT_CUTOFF_PASSED',
  INVALID_DATE = 'INVALID_DATE',
  INVALID_SLOT = 'INVALID_SLOT',
  MEAL_NOT_FOUND = 'MEAL_NOT_FOUND',
  MEAL_UNAVAILABLE = 'MEAL_UNAVAILABLE',
  SAME_MEAL_SELECTED = 'SAME_MEAL_SELECTED',
  DESTINATION_CONFLICT = 'DESTINATION_CONFLICT',
  SAME_DESTINATION_SELECTED = 'SAME_DESTINATION_SELECTED',
  ALREADY_SKIPPED = 'ALREADY_SKIPPED',
  NOTHING_TO_UNDO = 'NOTHING_TO_UNDO',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
}